requirements.txt
authlib==1.2.0
fastapi>=0.95.0
httpx>=0.25.1
itsdangerous>=2.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
redis>=4.5.1
requests>=2.31.0
starlette>=0.27.0
uvicorn[standard]>=0.22.0
